wallet: fix `gosec` linter errors on int32 conversion
We now properly fix the `int32 <-> uint32` conversion in the new code -
for the old code, it's safe to ignore cause they will be removed
eventually.

+++ wallet/utxos.go
@@ -33,7 +33,10 @@ type OutputSelectionPolicy struct {
 }
 
 func (p *OutputSelectionPolicy) meetsRequiredConfs(txHeight, curHeight int32) bool {
-	return confirmed(p.RequiredConfirmations, txHeight, curHeight)
+	return confirmed(
+		//nolint:gosec
+		uint32(p.RequiredConfirmations), txHeight, curHeight,
+	)
 }
 
 // UnspentOutputs fetches all unspent outputs from the wallet that match rules
